feat:add validate APIs
