+++ interface/Cargo.toml
@@ -18,6 +18,7 @@ wasm-bindgen = "0.2"
 
 chain-common = { path = "../chain-common" }
 wallet = { path = "../wallet" }
+crypto = { path = "../crypto" }
 
 [lib]
 name = "interface"
