+++ interface/src/handler.rs
@@ -1,596 +1,40 @@
-use super::coins::get_coin_info;
-use super::response_util::*;
-use crate::encode_message;
+mod account;
+mod common;
+mod sign;
+mod stored_key;
+mod validate;
 use chain_common::api::mw_request::Request::*;
-use chain_common::api::mw_response::Response;
-use chain_common::api::*;
-use chain_common::api::{mw_request, MwResponse, MwResponseError};
-use chain_common::ethereum;
-use prost::Message;
-use std::convert::From;
-
-use wallet::stored_key::*;
+use chain_common::api::{mw_request, MwResponse};
 
 pub fn dispatch_request(request: mw_request::Request) -> MwResponse {
     match request {
-        ParamLoadStoredKey(param) => load_stored_keys(param),
-        ParamCreateStoredKey(param) => create_stored_key(param),
-        ParamImportPrivateKey(param) => create_stored_key_with_private_key(param),
-        ParamImportMnemonic(param) => create_stored_key_with_mnemonic(param),
-        ParamImportJson(param) => create_with_json(param),
-        ParamGetStoredKeyAccountCount(param) => get_stored_key_account_count(param),
-        ParamGetStoredKeyAccount(param) => get_store_key_account(param),
-        ParamGetStoredKeyAllAccounts(param) => get_stored_key_all_accounts(param),
-        ParamGetStoredKeyAccountsOfCoin(param) => get_stored_key_accounts_of_coin(param),
-        ParamCreateAcccountOfCoinAtPath(param) => create_stored_key_account_of_coin_at_path(param),
-        ParamRemoveAccountsOfCoin(param) => remove_stored_key_account_of_coin(param),
-        ParamRemoveAccountOfAddress(param) => remove_account_of_address(param),
-        ParamExportPrivateKey(param) => export_private_key(param),
-        ParamExportPrivateKeyOfPath(param) => export_private_key_of_path(param),
-        ParamExportMnemonic(param) => export_mnemonic(param),
-        ParamExportKeyStoreJsonOfAddress(param) => export_key_store_json_of_address(param),
-        ParamExportKeyStoreJsonOfPath(param) => export_key_store_json_of_path(param),
-        ParamUpdateKeyStorePassword(param) => update_key_store_password(param),
-        ParamUpdateKeyStoreName(param) => update_key_store_name(param),
-        ParamSignTransaction(param) => sign_transaction(param),
-        ParamGetVersion(_) => get_lib_version(),
-    }
-}
+        ParamGetVersion(_) => common::get_lib_version(),
+        ParamLoadStoredKey(param) => stored_key::load_stored_keys(param),
+        ParamCreateStoredKey(param) => stored_key::create_stored_key(param),
+        ParamImportPrivateKey(param) => stored_key::create_stored_key_with_private_key(param),
+        ParamImportMnemonic(param) => stored_key::create_stored_key_with_mnemonic(param),
+        ParamImportJson(param) => stored_key::create_stored_key_with_json(param),
+        ParamExportPrivateKey(param) => stored_key::export_private_key(param),
+        ParamExportPrivateKeyOfPath(param) => stored_key::export_private_key_of_path(param),
+        ParamExportMnemonic(param) => stored_key::export_mnemonic(param),
+        ParamExportKeyStoreJsonOfAddress(param) => {
+            stored_key::export_key_store_json_of_address(param)
+        }
+        ParamExportKeyStoreJsonOfPath(param) => stored_key::export_key_store_json_of_path(param),
+        ParamUpdateKeyStorePassword(param) => stored_key::update_key_store_password(param),
+        ParamUpdateKeyStoreName(param) => stored_key::update_key_store_name(param),
 
-use std::env;
-pub const VERSION: &str = include_str!(concat!(env!("OUT_DIR"), "/VERSION"));
+        ParamGetStoredKeyAccountCount(param) => account::get_stored_key_account_count(param),
+        ParamGetStoredKeyAccount(param) => account::get_store_key_account(param),
+        ParamGetStoredKeyAllAccounts(param) => account::get_stored_key_all_accounts(param),
+        ParamGetStoredKeyAccountsOfCoin(param) => account::get_stored_key_accounts_of_coin(param),
+        ParamCreateAcccountOfCoinAtPath(param) => {
+            account::create_stored_key_account_of_coin_at_path(param)
+        }
+        ParamRemoveAccountsOfCoin(param) => account::remove_stored_key_account_of_coin(param),
+        ParamRemoveAccountOfAddress(param) => account::remove_account_of_address(param),
+        ParamSignTransaction(param) => sign::sign_transaction(param),
 
-fn get_lib_version() -> MwResponse {
-    MwResponse {
-        response: Some(Response::RespGetVersion(GetVersionResp {
-            version: VERSION.to_owned(),
-        })),
-    }
-}
-
-fn load_stored_keys(param: LoadStoredKeyParam) -> MwResponse {
-    let stored_keys_result: Result<Vec<StoredKey>, _> = param
-        .data
-        .iter()
-        .map(|json| serde_json::from_slice(&json))
-        .collect();
-    match stored_keys_result {
-        Ok(stored_keys) => MwResponse {
-            response: Some(Response::RespLoadStoredKey(LoadStoredKeyResp {
-                stored_keys: stored_keys.into_iter().map(StoredKeyInfo::from).collect(),
-            })),
-        },
-        Err(_) => get_json_error_response(),
-    }
-}
-
-fn create_stored_key(param: CreateStoredKeyParam) -> MwResponse {
-    let stored_key: StoredKey =
-        match StoredKey::create_with_mnemonic_random(&param.name, &param.password) {
-            Ok(key) => key,
-            Err(error) => {
-                return get_error_response_by_error(error);
-            }
-        };
-    MwResponse {
-        response: Some(Response::RespCreateStoredKey(CreateStoredKeyResp {
-            stored_key: Some(StoredKeyInfo::from(stored_key)),
-        })),
-    }
-}
-
-fn create_stored_key_with_private_key(param: ImportPrivateStoredKeyParam) -> MwResponse {
-    let stored_key =
-        StoredKey::create_with_private_key(&param.name, &param.password, &param.private_key);
-    match stored_key {
-        Ok(key) => MwResponse {
-            response: Some(Response::RespImportPrivateKey(ImportPrivateStoredKeyResp {
-                stored_key: Some(StoredKeyInfo::from(key)),
-            })),
-        },
-        Err(error) => get_error_response_by_error(error),
-    }
-}
-
-fn create_stored_key_with_mnemonic(param: ImportMnemonicStoredKeyParam) -> MwResponse {
-    let stored_key: StoredKey =
-        match StoredKey::create_with_mnemonic(&param.name, &param.password, &param.mnemonic) {
-            Ok(key) => key,
-            Err(error) => {
-                return get_error_response_by_error(error);
-            }
-        };
-    MwResponse {
-        response: Some(Response::RespImportMnemonic(ImportMnemonicStoredKeyResp {
-            stored_key: Some(StoredKeyInfo::from(stored_key)),
-        })),
-    }
-}
-
-fn create_with_json(param: ImportJsonStoredKeyParam) -> MwResponse {
-    let coin_info = get_coin_info(param.coin);
-    let coin = match coin_info {
-        Some(coin_info) => coin_info,
-        None => {
-            return MwResponse {
-                response: Some(Response::Error(MwResponseError {
-                    error_code: "-1".to_owned(),
-                    error_msg: "Invalid Coin Type".to_owned(),
-                })),
-            };
-        }
-    };
-    let stored_key: StoredKey =
-        match StoredKey::create_with_json(&param.name, &param.password, &param.json, coin) {
-            Ok(key) => key,
-            Err(error) => {
-                return get_error_response_by_error(error);
-            }
-        };
-    MwResponse {
-        response: Some(Response::RespImportJson(ImportJsonStoredKeyResp {
-            stored_key: Some(StoredKeyInfo::from(stored_key)),
-        })),
-    }
-}
-
-fn get_stored_key_account_count(param: GetStoredKeyAccountCountParam) -> MwResponse {
-    let stored_key: StoredKey = match serde_json::from_slice(&param.data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    MwResponse {
-        response: Some(Response::RespGetStoredKeyAccountCount(
-            GetStoredKeyAccountCountResp {
-                count: stored_key.get_accounts_count(),
-            },
-        )),
-    }
-}
-
-fn get_store_key_account(param: GetStoredKeyAccountParam) -> MwResponse {
-    let stored_key: StoredKey = match serde_json::from_slice(&param.data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    let account = match stored_key.get_account(param.index) {
-        Ok(account) => account,
-        Err(error) => {
-            return get_error_response_by_error(error);
-        }
-    };
-    MwResponse {
-        response: Some(Response::RespGetStoredKeyAccount(GetStoredKeyAccountResp {
-            account: Some(StoredKeyAccountInfo::from(account)),
-        })),
-    }
-}
-
-fn get_stored_key_all_accounts(param: GetStoredKeyAllAccountParam) -> MwResponse {
-    let stored_key: StoredKey = match serde_json::from_slice(&param.data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    let accounts_info: Vec<StoredKeyAccountInfo> = stored_key
-        .get_all_accounts()
-        .iter()
-        .map(StoredKeyAccountInfo::from)
-        .collect();
-    MwResponse {
-        response: Some(Response::RespGetStoredKeyAllAccounts(
-            GetStoredKeyAllAccountResp {
-                accounts: accounts_info,
-            },
-        )),
-    }
-}
-
-fn get_stored_key_accounts_of_coin(param: GetStoredKeyAccountsOfCoinParam) -> MwResponse {
-    let coin_info = get_coin_info(param.coin);
-    let coin = match coin_info {
-        Some(coin_info) => coin_info,
-        None => {
-            return MwResponse {
-                response: Some(Response::Error(MwResponseError {
-                    error_code: "-1".to_owned(),
-                    error_msg: "Invalid Coin Type".to_owned(),
-                })),
-            };
-        }
-    };
-    let stored_key: StoredKey = match serde_json::from_slice(&param.stored_key_data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-
-    let accounts: Vec<StoredKeyAccountInfo> = stored_key
-        .get_accounts_of_coin(coin)
-        .iter()
-        .map(StoredKeyAccountInfo::from)
-        .collect();
-    MwResponse {
-        response: Some(Response::RespGetStoredKeyAccountsOfCoin(
-            GetStoredKeyAccountsOfCoinResp {
-                stored_key: Some(StoredKeyInfo::from(stored_key)),
-                accounts,
-            },
-        )),
-    }
-}
-
-fn create_stored_key_account_of_coin_at_path(
-    param: CreateStoredKeyNewAccountAtPathParam,
-) -> MwResponse {
-    let coin_info = get_coin_info(param.coin);
-    let coin = match coin_info {
-        Some(coin_info) => coin_info,
-        None => {
-            return MwResponse {
-                response: Some(Response::Error(MwResponseError {
-                    error_code: "-1".to_owned(),
-                    error_msg: "Invalid Coin Type".to_owned(),
-                })),
-            };
-        }
-    };
-    let mut stored_key: StoredKey = match serde_json::from_slice(&param.stored_key_data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    let account = match stored_key.add_new_account_of_coin_and_derivation_path_by_password(
-        coin.clone(),
-        &param.derivation_path,
-        &param.password,
-    ) {
-        Ok(account) => account,
-        Err(error) => {
-            return get_error_response_by_error(error);
-        }
-    };
-    MwResponse {
-        response: Some(Response::RespCreateAccountOfCoinAtPath(
-            CreateStoredKeyNewAccountAtPathResp {
-                account: Some(StoredKeyAccountInfo::from(&account)),
-                stored_key: Some(StoredKeyInfo::from(stored_key)),
-            },
-        )),
-    }
-}
-
-fn remove_stored_key_account_of_coin(param: RemoveStoredKeyAccountOfCoinParam) -> MwResponse {
-    let coin_info = get_coin_info(param.coin);
-    let coin = match coin_info {
-        Some(coin_info) => coin_info,
-        None => {
-            return MwResponse {
-                response: Some(Response::Error(MwResponseError {
-                    error_code: "-1".to_owned(),
-                    error_msg: "Invalid Coin Type".to_owned(),
-                })),
-            };
-        }
-    };
-    let mut stored_key: StoredKey = match serde_json::from_slice(&param.stored_key_data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    stored_key.remove_accounts_of_coin(coin);
-    MwResponse {
-        response: Some(Response::RespRemoveAccountOfCoin(
-            RemoveStoredKeyAccountOfCoinResp {
-                stored_key: Some(StoredKeyInfo::from(stored_key)),
-            },
-        )),
-    }
-}
-
-fn remove_account_of_address(param: RemoveStoredKeyAccountOfAddressParam) -> MwResponse {
-    let coin_info = get_coin_info(param.coin);
-    let coin = match coin_info {
-        Some(coin_info) => coin_info,
-        None => {
-            return MwResponse {
-                response: Some(Response::Error(MwResponseError {
-                    error_code: "-1".to_owned(),
-                    error_msg: "Invalid Coin Type".to_owned(),
-                })),
-            };
-        }
-    };
-    let mut stored_key: StoredKey = match serde_json::from_slice(&param.stored_key_data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    match stored_key.remove_account_of_address(&param.address, coin) {
-        Ok(_) => MwResponse {
-            response: Some(Response::RespRemoveAccountOfAddress(
-                RemoveStoredKeyAccountOfAddressResp {
-                    stored_key: Some(StoredKeyInfo::from(stored_key)),
-                },
-            )),
-        },
-        Err(error) => get_error_response_by_error(error),
-    }
-}
-
-fn export_private_key(param: ExportKeyStorePrivateKeyParam) -> MwResponse {
-    let coin_info = get_coin_info(param.coin);
-    let coin = match coin_info {
-        Some(coin_info) => coin_info,
-        None => {
-            return MwResponse {
-                response: Some(Response::Error(MwResponseError {
-                    error_code: "-1".to_owned(),
-                    error_msg: "Invalid Coin Type".to_owned(),
-                })),
-            };
-        }
-    };
-    let mut stored_key: StoredKey = match serde_json::from_slice(&param.stored_key_data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    let private_key = match stored_key.export_private_key(&param.password, coin) {
-        Ok(key) => key,
-        Err(error) => {
-            return get_error_response_by_error(error);
-        }
-    };
-    MwResponse {
-        response: Some(Response::RespExportPrivateKey(
-            ExportKeyStorePrivateKeyResp { private_key },
-        )),
-    }
-}
-
-fn export_private_key_of_path(param: ExportKeyStorePrivateKeyOfPathParam) -> MwResponse {
-    let coin_info = get_coin_info(param.coin);
-    let coin = match coin_info {
-        Some(coin_info) => coin_info,
-        None => {
-            return MwResponse {
-                response: Some(Response::Error(MwResponseError {
-                    error_code: "-1".to_owned(),
-                    error_msg: "Invalid Coin Type".to_owned(),
-                })),
-            };
-        }
-    };
-    let mut stored_key: StoredKey = match serde_json::from_slice(&param.stored_key_data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    let private_key = match stored_key.export_private_key_of_path(
-        &param.password,
-        coin,
-        &param.derivation_path,
-    ) {
-        Ok(key) => key,
-        Err(error) => {
-            return get_error_response_by_error(error);
-        }
-    };
-    MwResponse {
-        response: Some(Response::RespExportPrivateKey(
-            ExportKeyStorePrivateKeyResp { private_key },
-        )),
-    }
-}
-
-fn export_mnemonic(param: ExportKeyStoreMnemonicParam) -> MwResponse {
-    let mut stored_key: StoredKey = match serde_json::from_slice(&param.stored_key_data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    let mnemonic = match stored_key.export_mnemonic(&param.password) {
-        Ok(key) => key,
-        Err(error) => {
-            return get_error_response_by_error(error);
-        }
-    };
-    MwResponse {
-        response: Some(Response::RespExportMnemonic(ExportKeyStoreMnemonicResp {
-            mnemonic,
-        })),
-    }
-}
-
-fn export_key_store_json_of_address(param: ExportKeyStoreJsonOfAddressParam) -> MwResponse {
-    let coin_info = get_coin_info(param.coin);
-    let coin = match coin_info {
-        Some(coin_info) => coin_info,
-        None => {
-            return MwResponse {
-                response: Some(Response::Error(MwResponseError {
-                    error_code: "-1".to_owned(),
-                    error_msg: "Invalid Coin Type".to_owned(),
-                })),
-            };
-        }
-    };
-    let mut stored_key: StoredKey = match serde_json::from_slice(&param.stored_key_data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    let json = match stored_key.export_key_store_json_of_address(
-        &param.password,
-        &param.new_password,
-        coin,
-        &param.address,
-    ) {
-        Ok(key) => key,
-        Err(error) => {
-            return get_error_response_by_error(error);
-        }
-    };
-    MwResponse {
-        response: Some(Response::RespExportKeyStoreJson(ExportKeyStoreJsonResp {
-            json,
-        })),
-    }
-}
-
-fn export_key_store_json_of_path(param: ExportKeyStoreJsonOfPathParam) -> MwResponse {
-    let coin_info = get_coin_info(param.coin);
-    let coin = match coin_info {
-        Some(coin_info) => coin_info,
-        None => {
-            return MwResponse {
-                response: Some(Response::Error(MwResponseError {
-                    error_code: "-1".to_owned(),
-                    error_msg: "Invalid Coin Type".to_owned(),
-                })),
-            };
-        }
-    };
-    let mut stored_key: StoredKey = match serde_json::from_slice(&param.stored_key_data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    let json = match stored_key.export_key_store_json_of_path(
-        &param.password,
-        &param.new_password,
-        coin,
-        &param.derivation_path,
-    ) {
-        Ok(key) => key,
-        Err(error) => {
-            return get_error_response_by_error(error);
-        }
-    };
-    MwResponse {
-        response: Some(Response::RespExportKeyStoreJson(ExportKeyStoreJsonResp {
-            json,
-        })),
-    }
-}
-
-fn update_key_store_password(param: UpdateStoredKeyPasswordParam) -> MwResponse {
-    let mut stored_key: StoredKey = match serde_json::from_slice(&param.stored_key_data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    match stored_key.update_password(&param.old_password, &param.new_password) {
-        Ok(_) => MwResponse {
-            response: Some(Response::RespUpdateKeyStorePassword(
-                UpdateStoredKeyPasswordResp {
-                    stored_key: Some(StoredKeyInfo::from(stored_key)),
-                },
-            )),
-        },
-        Err(error) => get_error_response_by_error(error),
-    }
-}
-
-fn update_key_store_name(param: UpdateStoredKeyNameParam) -> MwResponse {
-    let mut stored_key: StoredKey = match serde_json::from_slice(&param.stored_key_data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    stored_key.update_name(&param.new_name);
-    MwResponse {
-        response: Some(Response::RespUpdateKeyStoreName(UpdateStoredKeyNameResp {
-            stored_key: Some(StoredKeyInfo::from(stored_key)),
-        })),
-    }
-}
-
-fn sign_transaction(param: SignTransactionParam) -> MwResponse {
-    let coin_info = get_coin_info(param.coin);
-    let coin = match coin_info {
-        Some(coin_info) => coin_info,
-        None => {
-            return MwResponse {
-                response: Some(Response::Error(MwResponseError {
-                    error_code: "-1".to_owned(),
-                    error_msg: "Invalid Coin Type".to_owned(),
-                })),
-            };
-        }
-    };
-    let mut stored_key: StoredKey = match serde_json::from_slice(&param.stored_key_data) {
-        Ok(key) => key,
-        Err(_) => {
-            return get_json_error_response();
-        }
-    };
-    let input_struct = match param.input {
-        Some(input) => input,
-        None => {
-            return MwResponse {
-                response: Some(Response::Error(MwResponseError {
-                    error_code: "-1".to_owned(),
-                    error_msg: "Invalid sign input".to_owned(),
-                })),
-            };
-        }
-    };
-    let sign_transaction_param::Input::SignInput(chain_input) = input_struct;
-    let encoded_input = match encode_message(&chain_input) {
-        Ok(encoded) => encoded,
-        Err(_) => {
-            return MwResponse {
-                response: Some(Response::Error(MwResponseError {
-                    error_code: "-1".to_owned(),
-                    error_msg: "Invalid sign input".to_owned(),
-                })),
-            };
-        }
-    };
-    let sign_output = match stored_key.sign(&coin, &param.password, &param.address, &encoded_input)
-    {
-        Ok(key) => key,
-        Err(error) => {
-            return get_error_response_by_error(error);
-        }
-    };
-
-    let ethereum::SignInput { .. } = chain_input;
-    let decoded_output_result = ethereum::SignOutput::decode(&sign_output[..]);
-
-    let decoded_output = match decoded_output_result {
-        Ok(decoded) => decoded,
-        Err(_) => {
-            return MwResponse {
-                response: Some(Response::Error(MwResponseError {
-                    error_code: "-1".to_owned(),
-                    error_msg: "Invalid sign output".to_owned(),
-                })),
-            };
-        }
-    };
-
-    MwResponse {
-        response: Some(Response::RespSignTransaction(SignTransactionResp {
-            output: Some(sign_transaction_resp::Output::SignOutput(decoded_output)),
-        })),
+        ParamValidation(param) => validate::validate(param),
     }
 }
